Added javadocs to explain work to be done, as well as moved functionalit from App.js to Manager.js
